Extracts confirm delete from result forms
