+++ src/main/webapp/confirm-delete/index.js
@@ -1,0 +1,1 @@
+export { default } from './confirm-delete'
